+++ src/authfactory6.cpp
@@ -138,7 +138,7 @@ bool  AuthFactory::ip6_inc_counters(std::string& host, unsigned int rx, unsigned
     return ret;
 }
 
-// remove IP from AUTH IP MAP and synchronize with SHM AUTH IP TABLE (table which is used to communicate with bend daemon)
+// remove IP from AUTH IP MAP and synchronize with SHM AUTH IP TABLE (table which is used to communicate with bend service)
 void AuthFactory::ip6_remove(std::string& host) {
 
     std::scoped_lock<std::recursive_mutex> l_(AuthFactory::get_ip6_lock());
