+++ src/shm/shmauth.cpp
